Fix formatting port in test. Remove unnecessary group

+++ KaaTests/DefaultOperationTcpChannelTest.m
@@ -172,7 +172,7 @@
     [data appendBytes:&port length:sizeof(port)];
     ProtocolMetaData *md = [[ProtocolMetaData alloc] init];
     [md setConnectionInfo:data];
-    [md setAccessPointId:[[NSString stringWithFormat:@"%@:%lu", host, (unsigned long)port] hash]];
+    [md setAccessPointId:[[NSString stringWithFormat:@"%@:%i", host, port] hash]];
     [md setProtocolVersionInfo:pair];
     return md;
 }
